dashboard/pyproject.toml
[project]
name = "dashboard"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "duckdb==1.4.1",
    "pandas>=2.3.3",
    "plotly>=6.4.0",
    "prefect>=3.6.2",
    "prefect-client>=3.6.2",
    "python-dotenv>=1.2.1",
    "streamlit>=1.51.0",
]
